ingestion/requirements.txt
functions-framework
google-cloud-bigquery
requests
tenacity
pandas
db-dtypes